Move two NotificationsVC extensions to separate files
